Switched weather API used, added check for previous columns
Switched from Yahoo Weather API to Dark Sky API as Yahoo weather does not
tell information about chance of precipitation and Yahoo recently switched
their API usage policy, which also rendered the packaged used to be
broken for now.

The script now checks if the previous column is filled, else it fills it
now with the current travel time.